requirements.txt
# TUI Framework
textual>=0.47.0
rich>=13.7.0

# Stock Data - updated for Python 3.13 compatibility
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0

# Google APIs
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# System Monitoring
psutil>=5.9.6

# HTTP Requests
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0

# Terminal Charts
plotext>=5.2.8

# Development
pytest>=7.4.3
pytest-asyncio>=0.23.2
black>=23.12.0
ruff>=0.1.8
